Extract smoldot-light-base from wasm-node
