+++ bin/light-base/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "smoldot-light-wasm"
+name = "smoldot-light-base"
 version = "0.5.4"
 authors = ["Parity Technologies <admin@parity.io>", "Pierre Krieger <pierre.krieger1708@gmail.com>"]
 description = "Browser bindings to a light client for Substrate-based blockchains"
@@ -8,14 +8,16 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 edition = "2021"
 publish = false
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
 [dependencies]
+blake2-rfc = { version = "0.2.18", default-features = false }
+derive_more = "0.99.16"
+fnv = { version = "1.0.7", default-features = false }
 futures = "0.3.18"
-lazy_static = "1.4.0"
-log = { version = "0.4.14", features = ["std"] }
-pin-project = "1.0.8"
+hex = { version = "0.4.3", default-features = false }
+itertools = "0.10.1"
+log = { version = "0.4.14", default-features = false }
+lru = "0.7.0"
 rand = "0.8.4"
+serde_json = "1.0.72"
 slab = { version = "0.4.5", default-features = false }
-smoldot-light-base = { path = "../../light-base" }
+smoldot = { version = "0.2.0", path = "../..", default-features = false }
