+++ bin/light-base/src/lossy_channel.rs
