+++ bin/wasm-node/javascript/prepare.js
@@ -61,7 +61,7 @@ child_process.execSync(
 // The important step in this script is running `cargo build --target wasm32-wasi` on the Rust
 // code. This generates a `wasm` file in `target/wasm32-wasi`.
 child_process.execSync(
-    "cargo +" + rust_version + " build --package smoldot-light --target wasm32-wasi --no-default-features"
+    "cargo +" + rust_version + " build --package smoldot-light-wasm --target wasm32-wasi --no-default-features"
     + (build_profile == 'debug' ? '' : ' --' + build_profile),
     { 'stdio': 'inherit' }
 );
