+++ bin/wasm-node/rust/src/bindings.rs
